week-03/labs/lab-02-network-debugging/starter/api/requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
